added some investigation code make swarm and clusterRatelimiter testable less hardcoded fakeSwarm can now dynamically create localhost sockets fakeSwarm shares successfully data split the slidingwindow average algorithm into a testable function: data expectation or algorithm is wrong

+++ skipper.go
@@ -774,7 +774,7 @@ func Run(o Options) error {
 	var theSwarm *swarm.Swarm
 	if o.EnableSwarm {
 		swops := swarm.Options{
-			SwarmPort:        o.SwarmPort,
+			SwarmPort:        uint16(o.SwarmPort),
 			MaxMessageBuffer: o.SwarmMaxMessageBuffer,
 			LeaveTimeout:     o.SwarmLeaveTimeout,
 		}
